+++ Drivers/encoder.hpp
@@ -51,8 +51,8 @@ class Encoder {
     void activate();
     void deactivate();
     int32_t get_position();
-    double get_rpm();
-    double get_rad_per_sec();
+    float get_rpm();
+    float get_rad_per_sec();
     void reset();
 };
